pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["MarsBroshok96 <marsbroshok1996@gmail.com>"]
readme = "README.md"
packages = [{include = "page_analyzer"}]

[tool.poetry.dependencies]
python = "^3.10"
flask = "^2.2.2"
python-dotenv = "^0.21.0"
gunicorn = "^20.1.0"
validators = "^0.20.0"
psycopg2-binary = "^2.9.5"
requests = "^2.28.1"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
